devex: split up SkinPreviewrenderer
