+++ packages/ui/src/components/skin/SkinPreviewRenderer.vue
@@ -95,63 +95,25 @@
 
 <script setup lang="ts">
 import { ArrowLeftRightIcon, ClassicPlayerModel, SlimPlayerModel } from '@modrinth/assets'
-import {
-	applyCapeTexture,
-	applyTexture,
-	createTransparentTexture,
-	loadTexture as loadSkinTexture,
-} from '@modrinth/utils'
-import { useGLTF } from '@tresjs/cientos'
-import { TresCanvas, useRenderLoop, useTexture } from '@tresjs/core'
+import { TresCanvas } from '@tresjs/core'
 import * as THREE from 'three'
-import { clone as cloneSkeleton } from 'three/examples/jsm/utils/SkeletonUtils.js'
+import { computed, toRef, useSlots, useTemplateRef } from 'vue'
+
+import type {
+	SkinPreviewAnimationConfig,
+	SkinPreviewFitPadding,
+	SkinPreviewFraming,
+	SkinPreviewTuple,
+} from '#ui/composables/skin-rendering'
 import {
-	computed,
-	markRaw,
-	onBeforeMount,
-	onMounted,
-	onUnmounted,
-	ref,
-	shallowRef,
-	toRefs,
-	useSlots,
-	useTemplateRef,
-	watch,
-} from 'vue'
-
+	useSkinPreviewAnimation,
+	useSkinPreviewControls,
+	useSkinPreviewFit,
+	useSkinPreviewLoading,
+	useSkinPreviewScene,
+} from '#ui/composables/skin-rendering'
 import { useDynamicFontSize } from '../../composables'
-
-interface AnimationConfig {
-	baseAnimation: string
-	randomAnimations: string[]
-	randomAnimationInterval?: number
-	transitionDuration?: number
-}
-
-type SkinPreviewFraming = 'page' | 'modal'
-
-interface FitPadding {
-	top: number
-	right: number
-	bottom: number
-	left: number
-}
-
-interface FitLock {
-	containerSize: {
-		width: number
-		height: number
-	}
-	modelCenter: [number, number, number]
-	modelSize: [number, number, number]
-}
-
-type AnimationFinishedListener = (
-	event: THREE.AnimationMixerEventMap['finished'] & {
-		readonly type: 'finished'
-		readonly target: THREE.AnimationMixer
-	},
-) => void
+import { createRadialSpotlightShader } from './skin-preview-shader'
 
 const props = withDefaults(
 	defineProps<{
@@ -163,13 +125,13 @@ const props = withDefaults(
 		lockFit?: boolean
 		framing?: SkinPreviewFraming
 		fitZoom?: number
-		fitPadding?: Partial<FitPadding>
+		fitPadding?: Partial<SkinPreviewFitPadding>
 		/** @deprecated Manual framing fallback. */
 		scale?: number
 		/** @deprecated Manual framing fallback, or auto-fit FOV override when fit=true. */
 		fov?: number
 		initialRotation?: number
-		animationConfig?: AnimationConfig
+		animationConfig?: SkinPreviewAnimationConfig
 	}>(),
 	{
 		variant: 'CLASSIC',
@@ -192,20 +154,67 @@ const props = withDefaults(
 const skinPreviewContainer = useTemplateRef<HTMLElement>('skinPreviewContainer')
 const slots = useSlots()
 const nametagText = computed(() => props.nametag)
+const hasSubtitle = computed(() => Boolean(slots.subtitle))
+const hasNametagBadge = computed(() => Boolean(slots['nametag-badge']))
+const selectedModelSrc = computed(() =>
+	props.variant === 'SLIM' ? SlimPlayerModel : ClassicPlayerModel,
+)
 
-const fitEnabled = computed(() => {
-	if (props.fit !== undefined) return props.fit
-	return props.scale === undefined && props.fov === undefined
+const {
+	cleanupAnimationState,
+	clickImpulseOffsetX,
+	clickImpulseRotationZ,
+	clickImpulseScaleX,
+	clickImpulseScaleY,
+	currentAnimation,
+	getAvailableAnimations,
+	initializeAnimations,
+	playAnimation,
+	playClickInteraction,
+	stopAnimations,
+} = useSkinPreviewAnimation(toRef(props, 'animationConfig'))
+
+const { isModelLoaded, isTextureLoaded, modelCenter, modelSize, scene } = useSkinPreviewScene({
+	selectedModelSrc,
+	textureSrc: toRef(props, 'textureSrc'),
+	capeSrc: toRef(props, 'capeSrc'),
+	initializeAnimations,
+	cleanupAnimationState,
 })
-const currentFraming = computed<SkinPreviewFraming>(() => props.framing ?? 'page')
-const legacyScale = computed(() => props.scale ?? 1)
-const legacyFov = computed(() => props.fov ?? 40)
-const previewControlsPositionClass = computed(() =>
-	currentFraming.value === 'modal' ? 'bottom-[calc(6%)]' : 'bottom-[calc(15%+64px)]',
-)
-const subtitlePositionClass = computed(() =>
-	currentFraming.value === 'modal' ? 'bottom-[6%]' : 'bottom-[calc(15%)]',
-)
+
+const {
+	cameraConfig,
+	fitEnabled,
+	hasResolvedFit,
+	modelGroupPosition,
+	modelGroupScale,
+	modelOffset,
+	nametagTop,
+	previewControlsPositionClass,
+	spotlightPosition,
+	spotlightScale,
+	subtitlePositionClass,
+} = useSkinPreviewFit({
+	containerElement: computed(() => skinPreviewContainer.value),
+	fit: toRef(props, 'fit'),
+	lockFit: toRef(props, 'lockFit'),
+	framing: toRef(props, 'framing'),
+	fitZoom: toRef(props, 'fitZoom'),
+	fitPadding: toRef(props, 'fitPadding'),
+	scale: toRef(props, 'scale'),
+	fov: toRef(props, 'fov'),
+	nametag: toRef(props, 'nametag'),
+	hasSubtitle,
+	hasNametagBadge,
+	modelCenter,
+	modelSize,
+	isModelLoaded,
+})
+
+const rendererDpr: [number, number] = [1, 1.5]
+const radialSpotlightShader = createRadialSpotlightShader()
+const isReady = computed(() => isModelLoaded.value && isTextureLoaded.value && hasResolvedFit.value)
+const { isPreviewVisible, showLoading } = useSkinPreviewLoading(isReady)
 
 const { fontSize: nametagFontSize } = useDynamicFontSize({
 	containerElement: skinPreviewContainer,
@@ -217,964 +226,47 @@ const { fontSize: nametagFontSize } = useDynamicFontSize({
 	fontFamily: 'inherit',
 })
 
-const containerSize = ref({ width: 1, height: 1 })
-let resizeObserver: ResizeObserver | undefined
-
-onMounted(() => {
-	const el = skinPreviewContainer.value
-	if (!el) return
-
-	resizeObserver = new ResizeObserver(([entry]) => {
-		const { width, height } = entry.contentRect
-		containerSize.value = {
-			width: Math.max(width, 1),
-			height: Math.max(height, 1),
-		}
-
-		if (props.lockFit) {
-			lockFitState()
-		}
-	})
-
-	resizeObserver.observe(el)
-})
-
-const selectedModelSrc = computed(() =>
-	props.variant === 'SLIM' ? SlimPlayerModel : ClassicPlayerModel,
-)
-
-const scene = shallowRef<THREE.Object3D | null>(null)
-const lastCapeSrc = ref<string | undefined>(undefined)
-const texture = shallowRef<THREE.Texture | null>(null)
-const capeTexture = shallowRef<THREE.Texture | null>(null)
-const transparentTexture = createTransparentTexture()
-const rendererDpr: [number, number] = [1, 1.5]
-const LOADING_INDICATOR_DELAY_MS = 200
-const LOADING_INDICATOR_MIN_MS = 250
-const modelCenter = ref<[number, number, number]>([0, 1, 0])
-const modelSize = ref<[number, number, number]>([1, 2, 1])
-
-const isModelLoaded = ref(false)
-const isTextureLoaded = ref(false)
-const fitLock = ref<FitLock | null>(null)
-const hasUsableFitSize = computed(
-	() => containerSize.value.width > 1 && containerSize.value.height > 1,
-)
-const hasResolvedFit = computed(
-	() => !fitEnabled.value || (props.lockFit ? fitLock.value !== null : hasUsableFitSize.value),
-)
-const isReady = computed(() => isModelLoaded.value && isTextureLoaded.value && hasResolvedFit.value)
-const showLoading = ref(false)
-const isPreviewVisible = computed(() => isReady.value && !showLoading.value)
-let loadingIndicatorDelayTimer: number | null = null
-let loadingIndicatorMinTimer: number | null = null
-let loadingIndicatorShownAt = 0
-
-const fitContainerSize = computed(() =>
-	props.lockFit ? (fitLock.value?.containerSize ?? containerSize.value) : containerSize.value,
-)
-const fitModelCenter = computed(() =>
-	props.lockFit ? (fitLock.value?.modelCenter ?? modelCenter.value) : modelCenter.value,
-)
-const fitModelSize = computed(() =>
-	props.lockFit ? (fitLock.value?.modelSize ?? modelSize.value) : modelSize.value,
-)
-
-const mixer = ref<THREE.AnimationMixer | null>(null)
-const actions = ref<Record<string, THREE.AnimationAction>>({})
-const clock = new THREE.Clock()
-const currentAnimation = ref<string>('')
-const randomAnimationTimer = ref<number | null>(null)
-const lastRandomAnimation = ref<string>('')
-const animationFinishedListeners: AnimationFinishedListener[] = []
-
-const INTERACT_ANIMATION_NAME = 'interact'
-// TODO: apply fix in future to model-level
-const INTERACT_VISIBLE_DURATION_SECONDS = 0.5
-const CLICK_IMPULSE_MAX_ENERGY = 5
-const CLICK_IMPULSE_ENERGY_PER_CLICK = 1
-const CLICK_IMPULSE_DECAY_PER_SECOND = 6
-const CLICK_IMPULSE_BASE_SPEED = 18
-const CLICK_IMPULSE_SPEED_BOOST = 7
-const CLICK_IMPULSE_OFFSET_X = 0.035
-const CLICK_IMPULSE_ROTATION_Z = 0.055
-const CLICK_IMPULSE_SCALE_X = 0.018
-const CLICK_IMPULSE_SCALE_Y = 0.025
-let modelLoadVersion = 0
-let isUnmounted = false
-
-const clickImpulseEnergy = ref(0)
-const clickImpulsePhase = ref(0)
-const clickImpulseOffsetX = ref(0)
-const clickImpulseRotationZ = ref(0)
-const clickImpulseScaleX = ref(1)
-const clickImpulseScaleY = ref(1)
-
-const radialSpotlightShader = computed(() => ({
-	uniforms: {
-		innerColor: { value: new THREE.Color(0x000000) },
-		outerColor: { value: new THREE.Color(0xffffff) },
-		innerOpacity: { value: 0.3 },
-		outerOpacity: { value: 0.0 },
-		falloffPower: { value: 1.2 },
-		shadowRadius: { value: 7 },
-	},
-	vertexShader: `
-    varying vec2 vUv;
-    void main() {
-      vUv = uv;
-      gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-    }
-  `,
-	fragmentShader: `
-    uniform vec3 innerColor;
-    uniform vec3 outerColor;
-    uniform float innerOpacity;
-    uniform float outerOpacity;
-    uniform float falloffPower;
-    uniform float shadowRadius;
-
-    varying vec2 vUv;
-
-    void main() {
-      vec2 center = vec2(0.5, 0.5);
-      float dist = distance(vUv, center) * 2.0;
-
-      // Create shadow in the center
-      float shadowFalloff = 1.0 - smoothstep(0.0, shadowRadius, dist);
-
-      // Create overall spotlight falloff
-      float spotlightFalloff = 1.0 - smoothstep(0.0, 1.0, pow(dist, falloffPower));
-
-      // Combine both effects
-      vec3 color = mix(outerColor, innerColor, shadowFalloff);
-      float opacity = mix(outerOpacity, innerOpacity * shadowFalloff, spotlightFalloff);
-
-      gl_FragColor = vec4(color, opacity);
-    }
-  `,
-	transparent: true,
-	depthWrite: false,
-	depthTest: false,
-}))
-
-const FRAMING_PRESETS = {
-	page: {
-		fov: 35,
-		zoom: 0.96,
-		padding: { top: 0.2, right: 0.14, bottom: 0.3, left: 0.14 },
-	},
-	modal: {
-		fov: 35,
-		zoom: 1,
-		padding: { top: 0.1, right: 0.1, bottom: 0.18, left: 0.1 },
-	},
-} satisfies Record<SkinPreviewFraming, { fov: number; zoom: number; padding: FitPadding }>
-
-const resolvedFitPadding = computed<FitPadding>(() => {
-	const preset = FRAMING_PRESETS[currentFraming.value].padding
-
-	return {
-		top: Math.max(preset.top, props.nametag ? (slots['nametag-badge'] ? 0.28 : 0.2) : 0),
-		right: preset.right,
-		bottom: Math.max(preset.bottom, slots.subtitle ? 0.28 : preset.bottom),
-		left: preset.left,
-		...(props.fitPadding ?? {}),
-	}
-})
-
-const modelOffset = computed<[number, number, number]>(() => {
-	if (!fitEnabled.value) return [0, 0, 0]
-
-	const [x, y, z] = fitModelCenter.value
-	return [-x, -y, -z]
-})
-
-const modelGroupPosition = computed<[number, number, number]>(() => {
-	if (fitEnabled.value) return [0, 0, 0]
-	return [0, -0.05 * legacyScale.value, 1.95]
-})
-
-const modelGroupScale = computed<[number, number, number]>(() => {
-	if (fitEnabled.value) return [1, 1, 1]
-
-	const scale = 0.8 * legacyScale.value
-	return [scale, scale, scale]
-})
-
-const animatedModelGroupRotation = computed<[number, number, number]>(() => [
-	0,
-	modelRotation.value,
-	clickImpulseRotationZ.value,
-])
-
-const animatedModelGroupPosition = computed<[number, number, number]>(() => {
-	const [x, y, z] = modelGroupPosition.value
-	return [x + clickImpulseOffsetX.value, y, z]
-})
-
-const animatedModelGroupScale = computed<[number, number, number]>(() => {
-	const [x, y, z] = modelGroupScale.value
-	return [x * clickImpulseScaleX.value, y * clickImpulseScaleY.value, z]
-})
-
-const fittedCamera = computed(() => {
-	const width = Math.max(fitContainerSize.value.width, 1)
-	const height = Math.max(fitContainerSize.value.height, 1)
-	const aspect = width / height
-	const preset = FRAMING_PRESETS[currentFraming.value]
-	const padding = resolvedFitPadding.value
-
-	const usableWidth = Math.max(width * (1 - padding.left - padding.right), 1)
-	const usableHeight = Math.max(height * (1 - padding.top - padding.bottom), 1)
-
-	const [sizeX, sizeY, sizeZ] = fitModelSize.value
-	const halfWidth = Math.sqrt((sizeX / 2) ** 2 + (sizeZ / 2) ** 2)
-	const halfHeight = sizeY / 2
-
-	const fov = props.fov ?? preset.fov
-	const verticalFov = THREE.MathUtils.degToRad(fov)
-	const horizontalFov = 2 * Math.atan(Math.tan(verticalFov / 2) * aspect)
-
-	const paddedHalfWidth = halfWidth * (width / usableWidth)
-	const paddedHalfHeight = halfHeight * (height / usableHeight)
-	const zoom = Math.max((props.fitZoom ?? 1) * preset.zoom, 0.01)
-
-	const distance =
-		Math.max(
-			paddedHalfHeight / Math.tan(verticalFov / 2),
-			paddedHalfWidth / Math.tan(horizontalFov / 2),
-		) / zoom
-
-	const visibleHalfHeight = distance * Math.tan(verticalFov / 2)
-	const targetY = -(padding.bottom - padding.top) * visibleHalfHeight
-
-	return {
-		fov,
-		position: [0, targetY, -distance] as [number, number, number],
-		target: [0, targetY, 0] as [number, number, number],
-	}
-})
-
-const cameraConfig = computed(() => {
-	if (fitEnabled.value) return fittedCamera.value
-
-	return {
-		fov: legacyFov.value,
-		position: [0, 1.5, -3.25] as [number, number, number],
-		target: modelCenter.value,
-	}
-})
-
-const nametagTop = computed(() => {
-	if (!fitEnabled.value) return '18%'
-
-	const [, sizeY] = fitModelSize.value
-	const { fov, position, target } = cameraConfig.value
-	const distance = Math.max(Math.abs(position[2] - target[2]), 0.001)
-	const verticalFov = THREE.MathUtils.degToRad(fov)
-	const modelTopY = sizeY / 2
-	const projectedY = (modelTopY - target[1]) / distance / Math.max(Math.tan(verticalFov / 2), 0.001)
-	const topPercent = THREE.MathUtils.clamp(((1 - projectedY) / 2) * 100, 8, 40)
-
-	return `${topPercent - 2}%`
-})
-
 const nametagStyle = computed(() => ({
 	fontSize: nametagFontSize.value,
 	top: nametagTop.value,
 	transform: fitEnabled.value ? 'translate(-50%, calc(-100% - 0.75rem))' : 'translateX(-50%)',
 }))
 
-const spotlightY = computed(() => {
-	if (!fitEnabled.value) return -0.1 * legacyScale.value
-
-	const [, sizeY] = fitModelSize.value
-	return -sizeY / 2 - 0.02
+const {
+	ignoreControlClick,
+	modelRotation,
+	onCanvasClick,
+	onPointerDown,
+	onPointerMove,
+	onPointerUp,
+} = useSkinPreviewControls({
+	initialRotation: toRef(props, 'initialRotation'),
+	onClickWithoutDrag: () => {
+		playClickInteraction()
+	},
 })
 
-const spotlightPosition = computed<[number, number, number]>(() => [
+const animatedModelGroupRotation = computed<SkinPreviewTuple>(() => [
 	0,
-	spotlightY.value,
-	fitEnabled.value ? 0 : 2,
+	modelRotation.value,
+	clickImpulseRotationZ.value,
 ])
 
-const spotlightScale = computed<[number, number, number]>(() => {
-	if (!fitEnabled.value) {
-		const scale = 0.75 * legacyScale.value
-		return [scale, scale, scale]
-	}
-
-	const [sizeX, , sizeZ] = fitModelSize.value
-	const radius = Math.max(sizeX, sizeZ, 1) * 0.8
-	return [radius, radius, radius]
+const animatedModelGroupPosition = computed<SkinPreviewTuple>(() => {
+	const [x, y, z] = modelGroupPosition.value
+	return [x + clickImpulseOffsetX.value, y, z]
 })
 
-const { baseAnimation, randomAnimations } = toRefs(props.animationConfig)
-
-function initializeAnimations(loadedScene: THREE.Object3D, clips: THREE.AnimationClip[]) {
-	if (!clips || clips.length === 0) {
-		console.warn('No animation clips found in the model')
-		return
-	}
-
-	mixer.value = new THREE.AnimationMixer(loadedScene)
-	clock.start()
-	actions.value = {}
-
-	clips.forEach((clip) => {
-		if (clip.name === INTERACT_ANIMATION_NAME) {
-			clip.duration = INTERACT_VISIBLE_DURATION_SECONDS
-		}
-
-		const action = mixer.value!.clipAction(clip)
-
-		action.setLoop(THREE.LoopOnce, 1)
-		action.clampWhenFinished = true
-		actions.value[clip.name] = action
-	})
-
-	if (baseAnimation.value && actions.value[baseAnimation.value]) {
-		actions.value[baseAnimation.value].setLoop(THREE.LoopRepeat, Infinity)
-		playAnimation(baseAnimation.value, true)
-		setupRandomAnimationLoop()
-	} else {
-		console.warn(`Base animation "${baseAnimation.value}" not found`)
-
-		const firstAnimationName = Object.keys(actions.value)[0]
-		if (firstAnimationName) {
-			actions.value[firstAnimationName].setLoop(THREE.LoopRepeat, Infinity)
-			playAnimation(firstAnimationName, true)
-		}
-	}
-}
-
-function playAnimation(name: string, immediate = false) {
-	if (!mixer.value || !actions.value[name]) {
-		console.warn(`Animation "${name}" not found!`)
-		return false
-	}
-
-	const action = actions.value[name]
-
-	if (currentAnimation.value === name && action.isRunning() && name !== baseAnimation.value) {
-		console.log(`Animation "${name}" is already running, ignoring request`)
-		return false
-	}
-
-	const transitionDuration = props.animationConfig.transitionDuration || 0.3
-
-	Object.entries(actions.value).forEach(([actionName, actionInstance]) => {
-		if (actionName !== name && actionInstance.isRunning()) {
-			actionInstance.fadeOut(transitionDuration)
-		}
-	})
-
-	action.reset()
-
-	if (name === baseAnimation.value) {
-		action.setLoop(THREE.LoopRepeat, Infinity)
-	} else {
-		action.setLoop(THREE.LoopOnce, 1)
-		action.clampWhenFinished = true
-
-		const onFinished: AnimationFinishedListener = (event) => {
-			if (event.action === action) {
-				removeAnimationFinishedListener(onFinished)
-				if (currentAnimation.value === name && baseAnimation.value) {
-					action.fadeOut(transitionDuration)
-					const baseAction = actions.value[baseAnimation.value]
-					baseAction.reset()
-					baseAction.fadeIn(transitionDuration)
-					baseAction.play()
-					currentAnimation.value = baseAnimation.value
-				}
-			}
-		}
-
-		addAnimationFinishedListener(onFinished)
-	}
-
-	if (immediate) {
-		action.setEffectiveWeight(1)
-	} else {
-		action.fadeIn(transitionDuration)
-	}
-	action.play()
-
-	if (immediate) {
-		mixer.value.update(0)
-	}
-
-	currentAnimation.value = name
-	return true
-}
-
-function setupRandomAnimationLoop() {
-	const interval = props.animationConfig.randomAnimationInterval || 10000
-
-	function scheduleNextAnimation() {
-		if (randomAnimationTimer.value) {
-			clearTimeout(randomAnimationTimer.value)
-		}
-
-		randomAnimationTimer.value = window.setTimeout(() => {
-			if (randomAnimations.value.length > 0 && currentAnimation.value === baseAnimation.value) {
-				const availableAnimations = randomAnimations.value.filter(
-					(anim) => anim !== lastRandomAnimation.value,
-				)
-
-				// If all animations have been used, reset and use the full list
-				const animationsToChooseFrom =
-					availableAnimations.length > 0 ? availableAnimations : randomAnimations.value
-
-				const randomIndex = Math.floor(Math.random() * animationsToChooseFrom.length)
-				const randomAnimationName = animationsToChooseFrom[randomIndex]
-
-				if (actions.value[randomAnimationName]) {
-					lastRandomAnimation.value = randomAnimationName
-					playRandomAnimation(randomAnimationName)
-				}
-			} else {
-				// If not in base animation, wait and try again
-				scheduleNextAnimation()
-			}
-		}, interval)
-	}
-
-	scheduleNextAnimation()
-}
-
-function playRandomAnimation(name: string) {
-	if (!mixer.value || !actions.value[name]) {
-		console.warn(`Animation "${name}" not found!`)
-		return
-	}
-
-	const action = actions.value[name]
-
-	if (currentAnimation.value === name && action.isRunning()) {
-		console.log(`Animation "${name}" is already running, ignoring request`)
-		return
-	}
-
-	const transitionDuration = props.animationConfig.transitionDuration || 0.3
-
-	if (baseAnimation.value && actions.value[baseAnimation.value].isRunning()) {
-		actions.value[baseAnimation.value].fadeOut(transitionDuration)
-	}
-
-	action.reset()
-	action.setLoop(THREE.LoopOnce, 1)
-	action.clampWhenFinished = true
-	action.setEffectiveTimeScale(1)
-	action.fadeIn(transitionDuration)
-	action.play()
-
-	currentAnimation.value = name
-
-	const onFinished: AnimationFinishedListener = (event) => {
-		if (event.action === action) {
-			removeAnimationFinishedListener(onFinished)
-			if (currentAnimation.value === name && baseAnimation.value) {
-				action.fadeOut(transitionDuration)
-				const baseAction = actions.value[baseAnimation.value]
-				baseAction.reset()
-				baseAction.setEffectiveTimeScale(1)
-				baseAction.fadeIn(transitionDuration)
-				baseAction.play()
-				currentAnimation.value = baseAnimation.value
-
-				// Schedule the next random animation after returning to base
-				setupRandomAnimationLoop()
-			}
-		}
-	}
-
-	addAnimationFinishedListener(onFinished)
-}
-
-function addClickImpulse() {
-	clickImpulseEnergy.value = Math.min(
-		CLICK_IMPULSE_MAX_ENERGY,
-		clickImpulseEnergy.value + CLICK_IMPULSE_ENERGY_PER_CLICK,
-	)
-}
-
-function updateClickImpulse(delta: number) {
-	const energy = Math.max(0, clickImpulseEnergy.value - CLICK_IMPULSE_DECAY_PER_SECOND * delta)
-	clickImpulseEnergy.value = energy
-
-	if (energy <= 0) {
-		clickImpulseOffsetX.value = 0
-		clickImpulseRotationZ.value = 0
-		clickImpulseScaleX.value = 1
-		clickImpulseScaleY.value = 1
-		return
-	}
-
-	const intensity = energy / CLICK_IMPULSE_MAX_ENERGY
-	clickImpulsePhase.value += delta * (CLICK_IMPULSE_BASE_SPEED + energy * CLICK_IMPULSE_SPEED_BOOST)
-
-	const shake = Math.sin(clickImpulsePhase.value) * intensity
-	const squash = Math.abs(Math.sin(clickImpulsePhase.value * 1.7)) * intensity
-
-	clickImpulseOffsetX.value = shake * CLICK_IMPULSE_OFFSET_X
-	clickImpulseRotationZ.value = shake * CLICK_IMPULSE_ROTATION_Z
-	clickImpulseScaleX.value = 1 + squash * CLICK_IMPULSE_SCALE_X
-	clickImpulseScaleY.value = 1 - squash * CLICK_IMPULSE_SCALE_Y
-}
-
-function stopAnimations() {
-	if (mixer.value) {
-		mixer.value.stopAllAction()
-	}
-	currentAnimation.value = ''
-}
-
-function getAvailableAnimations(): string[] {
-	return Object.keys(actions.value)
-}
-
-function clearRandomAnimationTimer() {
-	if (randomAnimationTimer.value) {
-		clearTimeout(randomAnimationTimer.value)
-		randomAnimationTimer.value = null
-	}
-}
-
-function clearLoadingIndicatorDelayTimer() {
-	if (loadingIndicatorDelayTimer !== null) {
-		clearTimeout(loadingIndicatorDelayTimer)
-		loadingIndicatorDelayTimer = null
-	}
-}
-
-function clearLoadingIndicatorMinTimer() {
-	if (loadingIndicatorMinTimer !== null) {
-		clearTimeout(loadingIndicatorMinTimer)
-		loadingIndicatorMinTimer = null
-	}
-}
-
-function hideLoadingIndicatorAfterMinimum() {
-	const visibleFor = Date.now() - loadingIndicatorShownAt
-	const remaining = LOADING_INDICATOR_MIN_MS - visibleFor
-
-	if (remaining <= 0) {
-		showLoading.value = false
-		return
-	}
-
-	loadingIndicatorMinTimer = window.setTimeout(() => {
-		showLoading.value = false
-		loadingIndicatorMinTimer = null
-	}, remaining)
-}
-
-function addAnimationFinishedListener(listener: AnimationFinishedListener) {
-	mixer.value?.addEventListener('finished', listener)
-	animationFinishedListeners.push(listener)
-}
-
-function removeAnimationFinishedListener(
-	listener: AnimationFinishedListener,
-	targetMixer = mixer.value,
-) {
-	targetMixer?.removeEventListener('finished', listener)
-
-	const index = animationFinishedListeners.indexOf(listener)
-	if (index !== -1) {
-		animationFinishedListeners.splice(index, 1)
-	}
-}
-
-function clearAnimationFinishedListeners(targetMixer = mixer.value) {
-	animationFinishedListeners.forEach((listener) => {
-		targetMixer?.removeEventListener('finished', listener)
-	})
-	animationFinishedListeners.length = 0
-}
-
-function cleanupAnimationState(root: THREE.Object3D | null) {
-	clearRandomAnimationTimer()
-
-	const currentMixer = mixer.value
-	if (currentMixer) {
-		clearAnimationFinishedListeners(currentMixer)
-		currentMixer.stopAllAction()
-
-		if (root) {
-			currentMixer.uncacheRoot(root)
-		}
-	}
-
-	mixer.value = null
-	actions.value = {}
-	currentAnimation.value = ''
-	lastRandomAnimation.value = ''
-}
-
-function disposeSceneMaterials(root: THREE.Object3D | null) {
-	if (!root) return
-
-	const materials = new Set<THREE.Material>()
-
-	root.traverse((object) => {
-		const mesh = object as THREE.Mesh
-		if (!mesh.isMesh || !mesh.material) return
-
-		const meshMaterials = Array.isArray(mesh.material) ? mesh.material : [mesh.material]
-		meshMaterials.forEach((material) => materials.add(material))
-	})
-
-	materials.forEach((material) => material.dispose())
-}
+const animatedModelGroupScale = computed<SkinPreviewTuple>(() => {
+	const [x, y, z] = modelGroupScale.value
+	return [x * clickImpulseScaleX.value, y * clickImpulseScaleY.value, z]
+})
 
 defineExpose({
 	playAnimation,
 	stopAnimations,
 	getAvailableAnimations,
 	getCurrentAnimation: () => currentAnimation.value,
-})
-
-const { onLoop } = useRenderLoop()
-onLoop(() => {
-	const delta = clock.getDelta()
-
-	if (mixer.value) {
-		mixer.value.update(delta)
-	}
-
-	updateClickImpulse(delta)
-})
-
-const SKIN_LAYER_DEPTH_OFFSET: Record<string, number> = {
-	Body_Layer: -4,
-	Left_Leg_Layer: -8,
-	Right_Leg_Layer: -8,
-	Left_Arm_Layer: -12,
-	Right_Arm_Layer: -12,
-	Hat_Layer: -16,
-}
-
-function configureSkinPreviewMesh(mesh: THREE.Mesh) {
-	const isSkinLayer = mesh.name.endsWith('_Layer')
-	const layerDepthOffset = SKIN_LAYER_DEPTH_OFFSET[mesh.name] ?? 0
-	mesh.renderOrder = isSkinLayer ? Math.abs(layerDepthOffset) : 0
-
-	const materials = Array.isArray(mesh.material) ? mesh.material : [mesh.material]
-	materials.forEach((material) => {
-		if (!(material instanceof THREE.MeshStandardMaterial) || material.name === 'cape') return
-
-		material.transparent = false
-		material.alphaTest = 0.1
-		material.depthTest = true
-		material.depthWrite = true
-		material.polygonOffset = isSkinLayer
-		material.polygonOffsetFactor = layerDepthOffset
-		material.polygonOffsetUnits = layerDepthOffset
-		material.needsUpdate = true
-	})
-}
-
-function cloneSceneForRenderer(source: THREE.Object3D) {
-	const cloned = cloneSkeleton(source)
-
-	cloned.traverse((object) => {
-		const mesh = object as THREE.Mesh
-		if (!mesh.isMesh || !mesh.material) return
-
-		mesh.material = Array.isArray(mesh.material)
-			? mesh.material.map((material) => material.clone())
-			: mesh.material.clone()
-
-		configureSkinPreviewMesh(mesh)
-	})
-
-	return markRaw(cloned)
-}
-
-function cloneModelTuple(tuple: [number, number, number]): [number, number, number] {
-	return [tuple[0], tuple[1], tuple[2]]
-}
-
-function lockFitState() {
-	if (!fitEnabled.value || !props.lockFit || fitLock.value || !isModelLoaded.value) return
-
-	const { width, height } = containerSize.value
-	if (width <= 1 || height <= 1) return
-
-	fitLock.value = {
-		containerSize: { width, height },
-		modelCenter: cloneModelTuple(modelCenter.value),
-		modelSize: cloneModelTuple(modelSize.value),
-	}
-}
-
-async function loadModel(src: string) {
-	const loadVersion = ++modelLoadVersion
-
-	try {
-		isModelLoaded.value = false
-		fitLock.value = null
-		const { scene: loadedScene, animations } = await useGLTF(src)
-		const clonedScene = cloneSceneForRenderer(loadedScene)
-		if (isUnmounted || loadVersion !== modelLoadVersion) {
-			disposeSceneMaterials(clonedScene)
-			return
-		}
-
-		const previousScene = scene.value
-		cleanupAnimationState(previousScene)
-		disposeSceneMaterials(previousScene)
-		scene.value = clonedScene
-
-		if (texture.value) {
-			applyTexture(clonedScene, texture.value)
-		}
-
-		applyCapeTexture(clonedScene, capeTexture.value, transparentTexture)
-
-		if (animations && animations.length > 0) {
-			initializeAnimations(clonedScene, animations)
-		}
-
-		updateModelInfo()
-		isModelLoaded.value = true
-		lockFitState()
-	} catch (error) {
-		console.error('Failed to load model:', error)
-		if (!isUnmounted && loadVersion === modelLoadVersion) {
-			isModelLoaded.value = false
-		}
-	}
-}
-
-async function loadAndApplyTexture(src: string) {
-	if (!src) return null
-
-	try {
-		try {
-			return await loadSkinTexture(src)
-		} catch {
-			const tex = await useTexture([src])
-			tex.colorSpace = THREE.SRGBColorSpace
-			tex.flipY = false
-			tex.magFilter = THREE.NearestFilter
-			tex.minFilter = THREE.NearestFilter
-			return tex
-		}
-	} catch (error) {
-		console.error('Failed to load texture:', error)
-		return null
-	}
-}
-
-async function loadAndApplyCapeTexture(src: string | undefined) {
-	if (src === lastCapeSrc.value) return
-
-	lastCapeSrc.value = src
-
-	if (src) {
-		capeTexture.value = await loadAndApplyTexture(src)
-	} else {
-		capeTexture.value = null
-	}
-
-	if (scene.value) {
-		applyCapeTexture(scene.value, capeTexture.value, transparentTexture)
-	}
-}
-
-function getVisibleMeshBox(root: THREE.Object3D): THREE.Box3 | null {
-	root.updateWorldMatrix(true, true)
-
-	const result = new THREE.Box3()
-	const meshBox = new THREE.Box3()
-	let found = false
-
-	root.traverse((object) => {
-		const mesh = object as THREE.Mesh
-		if (!mesh.isMesh || !mesh.geometry || mesh.visible === false) return
-
-		const materials = Array.isArray(mesh.material) ? mesh.material : [mesh.material]
-		if (materials.length && materials.every((material) => material.visible === false)) return
-
-		if (!mesh.geometry.boundingBox) {
-			mesh.geometry.computeBoundingBox()
-		}
-
-		if (!mesh.geometry.boundingBox) return
-
-		meshBox.copy(mesh.geometry.boundingBox).applyMatrix4(mesh.matrixWorld)
-		result.union(meshBox)
-		found = true
-	})
-
-	return found && !result.isEmpty() ? result.clone() : null
-}
-
-function updateModelInfo() {
-	const box = scene.value ? getVisibleMeshBox(scene.value) : null
-
-	if (!box) {
-		modelCenter.value = [0, 1, 0]
-		modelSize.value = [1, 2, 1]
-		return
-	}
-
-	const center = new THREE.Vector3()
-	const size = new THREE.Vector3()
-
-	box.getCenter(center)
-	box.getSize(size)
-
-	modelCenter.value = [center.x, center.y, center.z]
-	modelSize.value = [Math.max(size.x, 0.001), Math.max(size.y, 0.001), Math.max(size.z, 0.001)]
-}
-
-const modelRotation = ref(props.initialRotation + Math.PI)
-const isDragging = ref(false)
-const previousX = ref(0)
-const hasDragged = ref(false)
-
-function onPointerDown(event: PointerEvent) {
-	;(event.currentTarget as HTMLElement).setPointerCapture(event.pointerId)
-	isDragging.value = true
-	previousX.value = event.clientX
-	hasDragged.value = false
-}
-
-function onPointerMove(event: PointerEvent) {
-	if (!isDragging.value) return
-	const deltaX = event.clientX - previousX.value
-	modelRotation.value += deltaX * 0.01
-	previousX.value = event.clientX
-	hasDragged.value = true
-}
-
-function onPointerUp(event: PointerEvent) {
-	isDragging.value = false
-	;(event.currentTarget as HTMLElement).releasePointerCapture(event.pointerId)
-}
-
-function onCanvasClick() {
-	if (!hasDragged.value) {
-		addClickImpulse()
-
-		if (actions.value[INTERACT_ANIMATION_NAME]) {
-			playRandomAnimation(INTERACT_ANIMATION_NAME)
-		}
-	}
-
-	hasDragged.value = false
-}
-
-function ignoreControlClick(event: MouseEvent) {
-	event.stopPropagation()
-}
-
-watch(
-	isReady,
-	(ready) => {
-		clearLoadingIndicatorDelayTimer()
-
-		if (ready) {
-			if (showLoading.value) {
-				clearLoadingIndicatorMinTimer()
-				hideLoadingIndicatorAfterMinimum()
-			}
-
-			return
-		}
-
-		clearLoadingIndicatorMinTimer()
-
-		if (showLoading.value || typeof window === 'undefined') {
-			return
-		}
-
-		loadingIndicatorDelayTimer = window.setTimeout(() => {
-			loadingIndicatorDelayTimer = null
-
-			if (isReady.value) {
-				return
-			}
-
-			showLoading.value = true
-			loadingIndicatorShownAt = Date.now()
-		}, LOADING_INDICATOR_DELAY_MS)
-	},
-	{ immediate: true },
-)
-
-watch(selectedModelSrc, (src) => loadModel(src))
-watch(
-	() => props.lockFit,
-	() => {
-		fitLock.value = null
-		lockFitState()
-	},
-)
-watch(
-	() => props.textureSrc,
-	async (newSrc) => {
-		isTextureLoaded.value = false
-		texture.value = await loadAndApplyTexture(newSrc)
-		if (scene.value && texture.value) {
-			applyTexture(scene.value, texture.value)
-		}
-		isTextureLoaded.value = true
-	},
-)
-watch(
-	() => props.capeSrc,
-	async (newCapeSrc) => {
-		await loadAndApplyCapeTexture(newCapeSrc)
-	},
-)
-
-watch(
-	() => props.animationConfig,
-	(newConfig) => {
-		clearRandomAnimationTimer()
-
-		if (mixer.value && newConfig.baseAnimation && actions.value[newConfig.baseAnimation]) {
-			playAnimation(newConfig.baseAnimation)
-			setupRandomAnimationLoop()
-		}
-	},
-	{ deep: true },
-)
-
-onBeforeMount(async () => {
-	try {
-		isTextureLoaded.value = false
-		texture.value = await loadAndApplyTexture(props.textureSrc)
-		isTextureLoaded.value = true
-
-		await loadModel(selectedModelSrc.value)
-
-		if (props.capeSrc) {
-			await loadAndApplyCapeTexture(props.capeSrc)
-		}
-	} catch (error) {
-		console.error('Failed to initialize skin preview:', error)
-	}
-})
-
-onUnmounted(() => {
-	isUnmounted = true
-	modelLoadVersion++
-	resizeObserver?.disconnect()
-	clearLoadingIndicatorDelayTimer()
-	clearLoadingIndicatorMinTimer()
-
-	cleanupAnimationState(scene.value)
-	disposeSceneMaterials(scene.value)
-	scene.value = null
-	transparentTexture.dispose()
 })
 </script>
