+++ packages/ui/src/utils/index.ts
@@ -13,3 +13,4 @@ export * from './tag-messages'
 export * from './truncate'
 export * from './version-compatibility'
 export * from './vue-children'
+export * from './webgl/skin-rendering'
