+++ packages/utils/index.ts
@@ -4,7 +4,6 @@ export * from './licenses'
 export * from './parse'
 export * from './projects'
 export * from './servers'
-export * from './three/skin-rendering'
 export * from './types'
 export * from './users'
 export * from './utils'
